+++ src/parse.c
@@ -11430,7 +11430,7 @@ static psl_node_t p_psl_low_bound(tree_t head)
 
    BEGIN_WITH_HEAD("PSL Low Bound", head);
 
-   return p_hdl_expression(head, PSL_TYPE_NUMERIC);
+   return p_hdl_expression(head, PSL_TYPE_NUMBER);
 }
 
 static psl_node_t p_psl_high_bound(void)
@@ -11446,13 +11446,13 @@ static psl_node_t p_psl_high_bound(void)
       tree_set_ival(inf, INT32_MAX);
 
       psl_node_t p = psl_new(P_HDL_EXPR);
-      psl_set_type(p, PSL_TYPE_NUMERIC);
+      psl_set_type(p, PSL_TYPE_NUMBER);
       psl_set_tree(p, inf);
       psl_set_loc(p, CURRENT_LOC);
       return p;
    }
    else
-      return p_hdl_expression(NULL, PSL_TYPE_NUMERIC);
+      return p_hdl_expression(NULL, PSL_TYPE_NUMBER);
 }
 
 static psl_node_t p_psl_range(tree_t head)
@@ -11698,7 +11698,7 @@ static psl_node_t p_psl_builtin_function_call(void)
          psl_add_operand(p, p1);
 
          if (optional(tCOMMA)) {
-            psl_node_t p2 = p_hdl_expression(NULL, PSL_TYPE_NUMERIC);
+            psl_node_t p2 = p_hdl_expression(NULL, PSL_TYPE_NUMBER);
             psl_add_operand(p, p2);
 
             if (optional(tCOMMA))
@@ -11745,7 +11745,7 @@ static psl_node_t p_psl_builtin_function_call(void)
 
    case PSL_BUILTIN_NONDET_VECTOR:
       {
-         psl_node_t p1 = p_hdl_expression(NULL, PSL_TYPE_NUMERIC);
+         psl_node_t p1 = p_hdl_expression(NULL, PSL_TYPE_NUMBER);
          psl_add_operand(p, p1);
 
          consume(tCOMMA);
@@ -11859,7 +11859,7 @@ static psl_node_t p_psl_count(void)
    if (peek() == tTO)
       return p_psl_range(count);
    else
-      return p_hdl_expression(count, PSL_TYPE_NUMERIC);
+      return p_hdl_expression(count, PSL_TYPE_NUMBER);
 }
 
 static psl_node_t p_psl_proc_block(psl_node_t value)
@@ -12001,7 +12001,7 @@ static void p_psl_formal_parameter(psl_node_t node)
    type_t type = p_psl_param_spec(node, &psl_type, &class);
 
    do {
-      tree_t p = tree_new(T_PARAM_DECL);
+      tree_t p = tree_new(T_PSL_PARAM_DECL);
       tree_set_ident(p, p_identifier());
       tree_set_loc(p, CURRENT_LOC);
       tree_set_class(p, class);
@@ -12024,7 +12024,7 @@ static void p_psl_formal_parameter_list(psl_node_t node)
       p_psl_formal_parameter(node);
 }
 
-static void p_psl_actual_parameter(psl_node_t node, bool seq)
+static void p_psl_actual_parameter(psl_node_t node, psl_type_t type)
 {
    // Actual_Parameter ::=
    //    AnyType | Sequence | Property
@@ -12033,12 +12033,32 @@ static void p_psl_actual_parameter(psl_node_t node, bool seq)
 
    BEGIN("PSL Actual parameter");
 
-   // "Sequence" includes "Any_Type" parsing (HDL or PSL expression)
-   // "Property" includes "Sequence" parsing
-   psl_add_operand(node, (seq) ? p_psl_sequence() : p_psl_property());
+   psl_node_t arg;
+   switch (type) {
+   case PSL_TYPE_BIT:
+   case PSL_TYPE_BOOLEAN:
+   case PSL_TYPE_BITVECTOR:
+   case PSL_TYPE_NUMERIC:
+   case PSL_TYPE_STRING:
+   case PSL_TYPE_HDLTYPE:
+      arg = p_hdl_expression(NULL, type);
+      break;
+   case PSL_TYPE_SEQUENCE:
+      arg = p_psl_sequence();
+      break;
+   case PSL_TYPE_ANY:
+   case PSL_TYPE_PROPERTY:
+      arg = p_psl_property();
+      break;
+   default:
+      fatal_trace("unexpected psl type kind %s in p_psl_actual_parameter",
+                  psl_type_str(type));
+   }
+
+   psl_add_operand(node, arg);
 }
 
-static void p_psl_actual_parameter_list(psl_node_t node, bool seq)
+static void p_psl_actual_parameter_list(psl_node_t p)
 {
 
    // sequence_Actual_Parameter { , sequence_Actual_Parameter }
@@ -12046,10 +12066,17 @@ static void p_psl_actual_parameter_list(psl_node_t node, bool seq)
 
    BEGIN("PSL Actual parameter list");
 
-   p_psl_actual_parameter(node, seq);
+   psl_node_t decl = psl_ref(p);
+   int decl_params = psl_ports(decl);
+   int n = 0;
 
-   while (optional(tCOMMA))
-      p_psl_actual_parameter(node, seq);
+   do {
+      tree_t param_decl = (n < decl_params) ? psl_port(decl, n) : NULL;
+      psl_type_t type = (param_decl) ? tree_subkind(param_decl) : PSL_TYPE_ANY;
+      p_psl_actual_parameter(p, type);
+      n++;
+   } while (optional(tCOMMA));
+
 }
 
 static psl_node_t p_psl_clocked_sere(psl_node_t head)
@@ -12306,7 +12333,7 @@ static psl_node_t p_psl_repeated_sere(psl_node_t head)
    return p;
 }
 
-static psl_node_t p_sequence_instance(tree_t decl)
+static psl_node_t p_psl_sequence_instance(tree_t decl)
 {
    // Name [ ( Actual_Parameter_List ) ]
 
@@ -12322,7 +12349,7 @@ static psl_node_t p_sequence_instance(tree_t decl)
    psl_set_ref(p, s_decl);
 
    if (optional(tLPAREN)) {
-      p_psl_actual_parameter_list(p, true);
+      p_psl_actual_parameter_list(p);
       consume(tRPAREN);
    }
 
@@ -12349,7 +12376,7 @@ static psl_node_t p_psl_sequence(void)
          if (tree_kind(name) == T_REF && tree_has_ref(name)) {
             tree_t decl = tree_ref(name);
             if (tree_kind(decl) == T_PSL_DECL)
-               head = p_sequence_instance(decl);
+               head = p_psl_sequence_instance(decl);
          }
 
          if (head == NULL)
@@ -12464,7 +12491,7 @@ static psl_node_t p_psl_fl_property(void)
             psl_set_flag(p, PSL_F_STRONG);
 
          if (optional(tLSQUARE)) {
-            psl_node_t count = p_hdl_expression(NULL, PSL_TYPE_NUMERIC);
+            psl_node_t count = p_hdl_expression(NULL, PSL_TYPE_NUMBER);
             psl_set_delay(p, count);
 
             consume(tRSQUARE);
@@ -12532,7 +12559,7 @@ static psl_node_t p_psl_fl_property(void)
          consume(tRPAREN);
 
          if (optional(tLSQUARE)) {
-            psl_node_t count = p_hdl_expression(NULL, PSL_TYPE_NUMERIC);
+            psl_node_t count = p_hdl_expression(NULL, PSL_TYPE_NUMBER);
             psl_set_delay(p, count);
 
             consume(tRSQUARE);